Fix broken time sensitive tests.
The time between creating the metadata and writing it to file may
exceed 1 second, which makes the test fail since the expected time
in_a_while(hours=valid_for) is created after that.
The expected timestamp should be created before to ensure the slow I/O
stuff doesn't interfere with the test result.