Fix for no Format in NameIDPolicy for SAML2 frontend
It is acceptable SAML2 to have an <AuthnRequest> with
a <NameIDPolicy> that includes the AllowCreate attribute but
that does not include the Format attribute. When the SAML2
frontend receives such a request it should default to the
metadata for the SP to determine which Format the SP requires.
Before this commit SATOSA would ignore the SP metadata in such
a case and use a transient format.