Select preferred binding on the configuration
pysaml2 supports the configuration option 'preferred_binding' which is
a map of services to array of preferred bindings for that service. Use
the configuration instead of the currently hardcoded choices.

An example configuration would be the following:

  module: satosa.backends.saml2.SAMLBackend
  name: Saml2
  config:
    sp_config:
      preferred_binding:
        single_sign_on_service:
        - urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect
        - urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST

This defines the preferred bindings for the SSO service, and is
equivalent the code that was hardcoded before. Changing the order of the
bindings would use HTTP-POST instead of the HTTP-Redirect.