Refactor determining primary identifier from IdP for LDAP query
Refactored how the attributes and NameID values asserted by the IdP
are processed to determine the identifier value to use when
constructing the LDAP filter value that is used to search for
the person record in LDAP.

The result is a breaking change in configuration syntax. The
configuration option idp_identifiers is no longer allowed and
is replaced by ordered_identifier_candidates.