Remember the IdP a user has selected from disco
This commit adds two new configuration variables that change the way
state is handled by SATOSA.

By default, SATOSA deletes the context state when it receives an
authentication response from an identity provider. The first configuration
option, CONTEXT_STATE_DELETE, allows us disable this behaviour and thus
keeping state across different authentication flows, when the user uses
the same browser.

The second configuration option, remember_selected_idp_from_disco,
controls whether SATOSA will remember and reuse the IdP that is returned
from a discovery service. If ForceAuthn is set in the authentication
request, then the user will then the user the IdP will be redirected to
the discovery service (if it is configured) and ForceAuthn will be set
in the authentication request towards the selected IdP.

These two options together allow us to modify the current behaviour so
that within a given session, a user will select only once the identity
provider and then SATOSA will store this information in the state
cookie. When the cookie expires, the user will be redirected again to
the discovery service.

TODO:

- Add tests