Configurable memoization of IdP selection when using MDQ
This commit introduces four optional configuration parameters that can be used
to modify the default SATOSA behaviour:

- CONTEXT_STATE_DELETE
- remember_selected_idp_from_disco
- use_disco_when_forceauthn
- mirror_saml_forceauthn

By default, SATOSA deletes the context state when it receives an
authentication response from an identity provider. The first configuration
option, CONTEXT_STATE_DELETE, allows us disable this behaviour and thus
keeping state across different authentication flows, when the user uses
the same browser.

The second configuration option, remember_selected_idp_from_disco,
controls whether SATOSA will remember and reuse the IdP that is returned
from a discovery service. If ForceAuthn is set in the authentication
request, then the user will then the user the IdP will be redirected to
the discovery service (if it is configured) and ForceAuthn will be set
in the authentication request towards the selected IdP.

These two options together allow us to modify the current behaviour so
that within a given session, a user will select only once the identity
provider and then SATOSA will store this information in the state
cookie. When the cookie expires, the user will be redirected again to
the discovery service.

The third configuration option, use_disco_when_forceauthn, controls
whether SATOSA will redirect the use to the discovery service, even when
remember_selected_idp_from_disco is true and for the current session
there is the entity id of an IdP stored in the cookie state. This
behaviour provides a way for SPs that need to force a new IdP selection
(e.g. for account linking purposes) to use ForceAuthn to achieve this.

The fourth configuration option, mirror_saml_forceauthn, adds configuration
option to mirror ForceAuthn. By default, when the SATOSA SAML frontend
receives a SAML authentication request with ForceAuthn set to `True`, this
information is not mirrored in the SAML authentication request that is
generated by the SATOSA SAML backend towards the upstream identity provider.
If the configuration parameter `mirror_saml_forceauthn` is set to `True`,
then the default behaviour changes and the SATOSA SAML backend will set
ForceAuthn to true when it proxies a SAML authentication request with
ForceAuthn set to `True`.

The default values of these configuration options are tuned so that the
default behaviour of SATOSA is not changed.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>