Support SAML NameID emailAddress and unspecified in response
Add logic that allows support for SAML NameID of types
urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress
urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified
urn:oasis:names:tc:SAML:2.0:nameid-format:unspecified
to be sent in the authentication response sent to SAML SPs.

This commit does not add any logic to help set the value for the
NameID but if it is set, say by a microservice, and if the metadata
for the SP specifies one of those types or one of those types is
specified by the SP in the authentication request, then this commit
allows the value to make it through and be set in the response.