Store NameID as subject_id from authn-requests
The Subject/NameID text value is stored in the internal data structure
as subject_id.  It can be used to forward an authn-request with the same
Subject element.

The Subject/NameID format is also available, but it remains to be
discussed how it should be handle in relation to the specified (name_id)
policy or the metadata settings.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>