Redo the configurable memoization capability
This commit introduces four optional configuration parameters that can
be used to modify the default SATOSA behaviour:

- CONTEXT_STATE_DELETE
- memorize_disco_idp
- use_memorized_disco_idp_when_force_authn
- mirror_saml_force_authn

By default, SATOSA deletes the context state when it receives an
authentication response from an identity provider. The first
configuration option, CONTEXT_STATE_DELETE, allows us disable this
behaviour and thus keep state across different authentication flows,
while the user uses the same browser.

The second configuration option, memorize_disco_idp, controls whether
SATOSA will remember and reuse the IdP that is selected from a discovery
service. If ForceAuthn is set in the authentication request, then the
user will be redirected to the discovery service (if it is configured)
and ForceAuthn will be set in the authentication request towards the
selected IdP.

These two options together allow us to modify the current behaviour so
that within a given session, a user will select only once the identity
provider and then SATOSA will store this information in the state
cookie. When the cookie expires, the user will be redirected again to
the discovery service.

The third configuration option,
use_memorized_disco_idp_when_force_authn, controls whether SATOSA will
skip the discovery service, even when memorize_disco_idp is set, for the
current session there is an entity id of an IdP stored in the cookie
state, and ForceAuthn is requested.

SPs that need to force a new IdP selection (e.g. for account linking
purposes) should set this option to False, in order to be able to use
ForceAuthn to redirect the user to the discovery service.

The fourth configuration option, mirror_saml_force_authn, adds
configuration option to mirror ForceAuthn. By default, when the SATOSA
SAML frontend receives a SAML authentication request with ForceAuthn set
to `True`, this information is not mirrored in the SAML authentication
request that is generated by the SATOSA SAML backend towards the
upstream identity provider. If the configuration parameter
`mirror_saml_force_authn` is set to `True`, then the default behaviour
changes and the SATOSA SAML backend will set ForceAuthn to true when it
proxies a SAML authentication request with ForceAuthn set to `True`.

The default values of these configuration options are tuned so that the
default behaviour of SATOSA is not changed.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>