Calculate force_authn value with correct type
The force_authn value comes from different places and could be either a
string or a boolean or NoneType. The value will later be used to set the
ForceAuthn property and it is expected to be a string.

We return 'true' to set the property, or False to omit it.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>