Adds 3 new configuration options for the Facebook backend
The Facebook backend does not request any scopes from Facebook. As a
result, the default scope (public_profile) is assumed and the user is
asked to give permission only for that. This has the effect that even
though in the query to the Facebook graph API we might be asking for the
user's e-mail, the graph API will return only the user data that are
part of the public profile.

Furthermore, if a user has given permission for some scopes, then the
the user response is remembered by Facebook and even if the application
changes its configuration and requests for more scopes, Facebook will
not take this into account, unless the 'auth_type' parameter is set to
'rerequest'

This commit introduces three new configuration options:

- 'auth_type' expects list of auth_types. If none is provided, then the
'auth_type' parameter will not be sent to the authorization_endpoint.

- 'scope' expects a list of scopes. If none is provided, then the scope
element will not be added and 'public_profile' will be assumed (as it
happens now)

- 'graph_endpoint' expects the endpoint for the Facebook graph API. Up
to now the value of the graph_endpoint was hardcoded in the actual code
of the backend. In order to retain backwards compatibility if the
configuration option for the 'graph_endpoint' is not set, then it is
assumed to have the value that used to be hardcoded.