Better default for auto_bind argument to ldap3.Connection object
A Python False is not an acceptable value for the auto_bind argument to
the ldap3.Connection object. This commit sets the default value to a module
defined constant that makes the most sense when trying to preserve
the REUSABLE strategy as the default (for now), and allows full
configuration by defining a mapping between configuration string values
and the ldap3 module constants, as is done for the client_strategy.