Better handling of single-value attribute by LdapAttributeStore
Better handling of single-value attributes from LDAP by the
LdapAttributeStore microservice. The ldap3 module response object
includes a raw_attributes dictionary and an attributes dictionary. The
attributes dictionary is populated with formatted values using
formatters based on standard schema syntaxes. If a particular attribute
in the schema is defined as multi valued, the attribute value in the
attributes dictionary is a list, but if the attribute in the schema is
not multi valued, the value in the attributes dictionary is a single
value.