Add mapping for OIDC multivalue claims
Core OIDC claims define the representation of their values when those
can have multiple values. We should respect those forms. This mapping
defines if and how the values should be combined.

For all other claims, we return the same values as the ones that have
been set in the internal representation.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>