Add KEY_METADATA_STORE to store the saml2 metadata store
KEY_METADATA_STORE will hold the metadata store for saml2 backends and
frontends. Accessing it that through the Context object will allow one
to peek into the metadata of the corresponding backend or frontend.

Previously, KEY_BACKEND_METADATA_STORE was used for this reason, but was
available only on the backend. It is now a deprecated alias.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>