Remove the metadata_construction param
Additionally, the saml2.assertion.Policy object can be initialized with
a metadata store and thus the .restrict and .filter methods do not need
such a param. This will remain as it was, until some time has passed and
confidence is built that peolpe are using a recent enough version of
pysaml2, before dropping the param from the .restrict method. Up to that
point, warnings will be output, but functionality is preserved.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>