fix: PrimaryIdentifier: fix clear_input_attributes
The clear_input_attributes functionality in the PrimaryIdentifier MicroService
wasn't working - the original code:

        if clear_input_attributes:
            msg = "{} Clearing values for these input attributes: {}".format(
                logprefix, data.attribute_names
            )
            logline = lu.LOG_FMT.format(id=lu.get_session_id(context.state), message=msg)
            logger.debug(logline)
            data.attributes = {}

* would break on the use of data.attribute_names (KeyError)
* would clear ALL attributes

I assume the intended functionality was to clear only the attributes used to construct the values
- so `candidate['attribute_names']` for the successful `candidate`.

To have access to the candidate object, this code needs to be moved into `constructPrimaryIdentifier`
- and in order to have access to the `clear_input_attributes` option value, this is added
as an optional parameter to `constructPrimaryIdentifier`.

And, as `'name_id'` would not be found among the attributes, exclude it from the list of attributes to clear.