Introduce explicit context property to hold the query params
A POST request can have a query string. In that case context.request
will hold the data from the POST request body, and thus there is no
place to hold the query params.

With this changeset, a new property is introduced to hold the query
string, parsed as query params. The query params is a list of tuples.
Each tuple holds two elements, the query param name and the query param
value. Params with no value are dropped.

    ?param_w_value=123&param_w_no_value

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>