Select LDAP config by extracted attribute
This patch introduces a new global config variable `provider_attribute` to make
it possible to select the config not only by entity but also select the config
variable by a previous set attribute.
This way it is possible to use a single point of authentication, but enrich the
information from different ldap server based on e.g. the domain attribute
extracted in previous steps.