squash! saml2 backend: support using multiple ACS URLs
Make ACS selection configurable with the `acs_selection_strategy`
parameter, keeping the default backwards-compatible (`use_first_acs`).
Added the relevant example and documentation.

Additionally, log an error (instead of debug) message if the
authentication request can not be constructed, since most of the time
this is a configuration or environment error.