saml2 backend: support using multiple ACS URLs (#409)
* saml2 backend: support using multiple ACS URLs

When Satosa sends out a SAML2 AuthnRequest, it specifies the
AssertionConsumerServiceUrl parameter as well, unless the
`hide_assertion_consumer_service` configuration parameter is set. However,
Satosa might be deployed in an environment where not all interfaces and host
names are accessible for all users. After this change, Satosa tries to
select the ACS URL based on the current request, and falls back to the first
ACS if there is no match.

* squash! saml2 backend: support using multiple ACS URLs

Make ACS selection configurable with the `acs_selection_strategy`
parameter, keeping the default backwards-compatible (`use_first_acs`).
Added the relevant example and documentation.

Additionally, log an error (instead of debug) message if the
authentication request can not be constructed, since most of the time
this is a configuration or environment error.