backends/saml2: gracefully handle mismatching ACS
When the IdP redirects to an ACS which has a different address than the
one we used for initiating the request, we are unable to verify the
RelayState, since the browser does not send the session cookie. In order
to make configuration debugging easier, raise an explanatory
SATOSAAuthenticationError instead of a KeyError.

While adding a unit test to check for the proper error reporting, some
code duplication was refactored.